requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
et-xmlfile==2.0.0
fastapi==0.116.1
flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
fpdf==1.7.2
fpdf2==2.8.3
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
keras==3.11.0
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.0
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.5.3
msgpack==1.1.1
namex==0.1.0
narwhals==2.0.0
numba==0.61.2
numpy==2.1.3
opencv-python==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.4.0
optree==0.17.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.54.0
plotly==6.2.0
pooch==1.8.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pydyf==0.11.0
pyee==13.0.0
pygments==2.19.2
pyhanko==0.29.1
pyhanko-certvalidator==0.27.0
pyparsing==3.2.3
pypdf==5.9.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
reportlab==4.4.3
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
starlette==0.47.2
svglib==1.5.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
